+++ Sources/Sentry/include/SentryProfilingSwiftHelpers.h
@@ -19,7 +19,7 @@ NS_ASSUME_NONNULL_BEGIN
 extern "C" {
 #endif
 
-BOOL sentry_isContinuousProfilingV2Enabled(SentryClient *client);
+BOOL sentry_isContinuousProfilingEnabled(SentryClient *client);
 BOOL sentry_isProfilingCorrelatedToTraces(SentryClient *client);
 SentryProfileOptions *_Nullable sentry_getProfiling(SentryClient *client);
 NSString *sentry_stringFromSentryID(SentryId *sentryID);